Introduce handler design pattern
- Remove extra config layer
- Initialize dependencies on starting up
- Re-org internal packages
- Decouple more dependencies
