+++ internal/storage/ddb.go
@@ -1,4 +1,3 @@
-// XXX: Currently transactional operations are not used.
 package storage
 
 import (
@@ -6,7 +5,6 @@ import (
 	"strconv"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
 	av "github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
@@ -23,22 +21,17 @@ type Record struct {
 	CreatedAt   string `dynamodbav:"created_at"`
 }
 
-type Storage struct {
-	client    *dynamodb.Client
+type DDB struct {
+	inner     *dynamodb.Client
 	tableName *string
 }
 
-func NewStorage(ctx context.Context, tableName string) (*Storage, error) {
-	cfg, err := config.LoadDefaultConfig(ctx)
-	if err != nil {
-		return &Storage{}, errors.Wrap(err, "failed to load AWS config")
-	}
-	client := dynamodb.NewFromConfig(cfg)
-
-	return &Storage{client: client, tableName: &tableName}, nil
+func NewDDB(ctx context.Context, awsConfig aws.Config, tableName string) (*DDB, error) {
+	inner := dynamodb.NewFromConfig(awsConfig)
+	return &DDB{inner: inner, tableName: &tableName}, nil
 }
 
-func (s *Storage) Save(ctx context.Context, rec Record) error {
+func (s *DDB) Save(ctx context.Context, rec Record) error {
 	m, err := av.MarshalMap(rec)
 	if err != nil {
 		return errors.Wrapf(err, "failed to marshal record: %+v", rec)
@@ -47,7 +40,7 @@ func (s *Storage) Save(ctx context.Context, rec Record) error {
 		Item:      m,
 		TableName: s.tableName,
 	}
-	if _, err := s.client.PutItem(ctx, &input); err != nil {
+	if _, err := s.inner.PutItem(ctx, &input); err != nil {
 		return errors.Wrap(err, "failed to put item")
 	}
 	return nil
@@ -55,14 +48,14 @@ func (s *Storage) Save(ctx context.Context, rec Record) error {
 
 // QueryByChannelName returns found Records sorted by .Version with descending order.
 // https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Query.html
-func (s *Storage) QueryByChannelName(ctx context.Context, channelName string) ([]Record, error) {
+func (s *DDB) QueryByChannelName(ctx context.Context, channelName string) ([]Record, error) {
 	input := dynamodb.QueryInput{
 		TableName:                 s.tableName,
 		KeyConditionExpression:    aws.String("channel_name = :channel_name"),
 		ExpressionAttributeValues: itemMap{":channel_name": &types.AttributeValueMemberS{Value: channelName}},
 		ScanIndexForward:          aws.Bool(true),
 	}
-	out, err := s.client.Query(ctx, &input)
+	out, err := s.inner.Query(ctx, &input)
 	if err != nil {
 		return []Record{}, errors.Wrap(err, "failed to query")
 	}
@@ -79,7 +72,7 @@ func (s *Storage) QueryByChannelName(ctx context.Context, channelName string) ([
 }
 
 // Delete removes a record. The record must be in the table.
-func (s *Storage) Delete(ctx context.Context, rec Record) error {
+func (s *DDB) Delete(ctx context.Context, rec Record) error {
 	input := dynamodb.DeleteItemInput{
 		TableName: s.tableName,
 		Key: itemMap{
@@ -91,7 +84,7 @@ func (s *Storage) Delete(ctx context.Context, rec Record) error {
 		ExpressionAttributeNames:  map[string]string{"#t": "token"},
 		ReturnValues:              types.ReturnValueAllOld,
 	}
-	out, err := s.client.DeleteItem(ctx, &input)
+	out, err := s.inner.DeleteItem(ctx, &input)
 	if err != nil {
 		return errors.Wrap(err, "failed to delete")
 	}
@@ -102,7 +95,7 @@ func (s *Storage) Delete(ctx context.Context, rec Record) error {
 	return nil
 }
 
-func (s *Storage) ScanAll(ctx context.Context) ([]Record, error) {
+func (s *DDB) ScanAll(ctx context.Context) ([]Record, error) {
 	var (
 		recs              []Record
 		exclusiveStartKey itemMap
@@ -113,7 +106,7 @@ func (s *Storage) ScanAll(ctx context.Context) ([]Record, error) {
 			TableName:         s.tableName,
 			ExclusiveStartKey: exclusiveStartKey,
 		}
-		out, err := s.client.Scan(ctx, &input)
+		out, err := s.inner.Scan(ctx, &input)
 		if err != nil {
 			return []Record{}, errors.Wrap(err, "failed to scan")
 		}
